Rename file to reflect purpse.
